0.15.49 - added recursive function to variable, allowing for better tracking of nested variables - The above prevents an infinite loop that was occcuring - CANNOT for the LIFE of me figure out what the two f*cking pixel gap in the variable is caused by - seems to occur when variable is moved after being attached to a gap, as if some scaling is being changed - doesnt occur for smaller or taller variables

+++ src/heightchanger.cpp
@@ -21,7 +21,7 @@ void HeightChanger::UpdateHeight(){
     height -= parts.top -> getCurrentFrame().h * parts.top -> getScale().y;
     height -= parts.bottom -> getCurrentFrame().h * parts.bottom -> getScale().y;
 
-    parts.left -> setScale(Vector2f(parts.left -> getScale().x, height / parts.left -> getCurrentFrame().h));
+    parts.left -> setScale(Vector2f(parts.left -> getScale().x, height / (float)parts.left -> getCurrentFrame().h));
     parts.center -> setScale(Vector2f(parts.center -> getScale().x, parts.left -> getScale().y));
     parts.right -> setScale(parts.left -> getScale());
